# users/config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "N/A",
        "title": "SCIM",
        "description": "The SCIM API lets you manage users in your organization. You can then automate the provisioning of product licenses for these users, and they can use your company's Single Sign-On solution through an Identity Provider. The SCIM API lets you manage users in your organization. You can then automate the provisioning of product licenses for these users, and they can use your company's Single Sign-On solution through an Identity Provider."
    },
    "host": "pldgicrgsdfkmt6iq-mock.stoplight-proxy.io",
    "basePath": "/v1",
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/Users/{userKey}": {
            "parameters": [
                {
                    "name": "userKey",
                    "in": "path",
                    "description": "The key of the user to query. The user must be in the organization domain",
                    "required": true,
                    "type": "integer",
                    "format": "int64"
                }
            ],
            "get": {
                "operationId": "getUser",
                "summary": "Get User",
                "tags": [
                    "Users"
                ],
                "description": "Queries user identity in the organization domain.",
                "responses": {
                    "200": {
                        "description": "The request has succeeded.",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "401": {
                        "description": "Client is not sufficiently authorized"
                    },
                    "403": {
                        "description": "Invalid token passed"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "502": {
                        "description": "Authentication or account gateway error occurred"
                    },
                    "504": {
                        "description": "Authentication or account gateway timeout occurred"
                    }
                }
            },
            "patch": {
                "operationId": "updateUser",
                "summary": "Update User",
                "tags": [
                    "Users"
                ],
                "description": "Changes a limited set (or all if you choose) of the user's data. The updated user email domain must be an existing organization email domain.",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/UserDefinition"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The user has been updated.",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "One of the following requirements is not met:<br />The userName field is required.<br />The userName field must be non-empty.<br />The userName field cannot exceed 128 characters.<br />The locale field must be composed of 1 or 2 parts.<br />The locale language sub-field must be ISO-639.<br />The locale country sub-field must be ISO-3166.<br />The timezone field must be a valid timezone.<br />The givenName field must not exceed 60 characters.<br />The familyName field must not exceed 60 characters."
                    },
                    "401": {
                        "description": "Client is not sufficiently authorized"
                    },
                    "403": {
                        "description": "Invalid token passed"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "409": {
                        "description": "Email address conflict"
                    },
                    "502": {
                        "description": "Authentication or account gateway error occurred"
                    },
                    "504": {
                        "description": "Authentication or account gateway timeout occurred"
                    }
                }
            },
            "delete": {
                "operationId": "deleteUser",
                "summary": "Delete User",
                "tags": [
                    "Users"
                ],
                "description": "Deletes a user from the organization (but not from the account).",
                "responses": {
                    "200": {
                        "description": "The request has succeeded."
                    },
                    "401": {
                        "description": "Client is not sufficiently authorized"
                    },
                    "403": {
                        "description": "Invalid token passed"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "502": {
                        "description": "Authentication or account gateway error occurred"
                    },
                    "504": {
                        "description": "Authentication or account gateway timeout occurred"
                    }
                }
            },
            "put": {
                "operationId": "replaceUser",
                "summary": "Replace User",
                "tags": [
                    "Users"
                ],
                "description": "Changes an existing user's displayName, locale, timezone, username and password. The request must include the full user definition (to modify one or more values without sending the full definition, see Update User). The replaced user email domain must be an existing organization email domain.",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/UserDefinition"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The user has been replaced.",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "One of the following requirements is not met:<br />The userName field is required.<br />The userName field must be non-empty.<br />The userName field cannot exceed 128 characters.<br />The locale field must be composed of 1 or 2 parts.<br />The locale language sub-field must be ISO-639.<br />The locale country sub-field must be ISO-3166.<br />The timezone field must be a valid timezone.<br />The givenName field must not exceed 60 characters.<br />The familyName field must not exceed 60 characters."
                    },
                    "401": {
                        "description": "Client is not sufficiently authorized"
                    },
                    "403": {
                        "description": "Invalid token passed"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "409": {
                        "description": "Email address conflict"
                    },
                    "502": {
                        "description": "Authentication or account gateway error occurred"
                    },
                    "504": {
                        "description": "Authentication or account gateway timeout occurred"
                    }
                }
            }
        },
...
    },
...
    "definitions": {
        "UserDefinition": {
            "title": "UserDefinition",
            "description": "Describes a new user",
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string",
                    "description": "The user's display name",
                    "x-faker": "internet.userName"
                },
                "locale": {
                    "description": "The user's language settings",
                    "type": "string"
                },
                "name": {
                    "$ref": "#/definitions/UserFullName"
                },
                "timezone": {
                    "description": "The user's time zone",
                    "type": "string"
                },
                "userName": {
                    "description": "The user's username, usually their email address",
                    "type": "string",
                    "x-faker": "internet.userName"
                },
                "emails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Email"
                    }
                },
                "phoneNumbers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/PhoneNumber"
                    }
                },
                "schemas": {
                    "type": "array",
                    "default": [],
                    "uniqueItems": true,
                    "items": {
                        "type": "string",
                        "enum": [
                            "urn:scim:schemas:core:1.0",
                            "urn:scim:schemas:extension:italia:1.0"
                        ]
                    }
                },
                "urn:scim:schemas:extension:italia:1.0": {
                    "type": "object",
                    "required": [
                        "fiscalNumber"
                    ],
                    "properties": {
                        "fiscalNumber": {
                            "type": "string"
                        },
                        "optOut": {
                            "type": "array",
                            "uniqueItems": true,
                            "items": {
                                "type": "object"
                            }
                        }
                    }
                }
            },
            "required": [
                "displayName",
                "name",
                "emails",
                "schemas",
                "urn:scim:schemas:extension:italia:1.0"
            ]
        },
...
        "UserMetadata": {
            "title": "UserMetadata",
            "description": "User metadata",
            "properties": {
                "created": {
                    "description": "The date and time the user was created",
                    "format": "date-time",
                    "type": "string"
                },
                "location": {
                    "description": "The URL where the user can be accessed",
                    "type": "string"
                }
            },
            "required": [
                "created",
                "location"
            ]
        },
        "UserFullName": {
            "title": "UserFullName",
            "description": "The full name of the user",
            "properties": {
                "familyName": {
                    "description": "The user's surname",
                    "type": "string",
                    "x-faker": "name.lastName"
                },
                "givenName": {
                    "description": "The user's first name",
                    "type": "string",
                    "x-faker": "name.firstName"
                }
            },
            "required": [
                "familyName",
                "givenName"
            ]
        },
        "User": {
            "title": "User",
            "description": "Describes a user belonging to an organization",
            "allOf": [
                {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string",
                            "description": "The user's unique key"
                        },
                        "meta": {
                            "$ref": "#/definitions/UserMetadata"
                        }
                    },
                    "required": [
                        "id"
                    ]
                },
                {
                    "$ref": "#/definitions/UserDefinition"
                }
            ]
        },
...
        "Email": {
            "title": "Email",
            "type": "object",
            "properties": {
                "type": {
                    "type": "string",
                    "x-faker": "lorem.word"
                },
                "primary": {
                    "type": "boolean"
                },
                "value": {
                    "type": "string",
                    "format": "email",
                    "x-faker": "internet.email"
                }
            },
            "required": [
                "value"
            ]
        },
        "PhoneNumber": {
            "title": "PhoneNumber",
            "type": "object",
            "properties": {
                "value": {
                    "type": "string",
                    "x-faker": "phone.phoneNumber"
                },
                "type": {
                    "type": "string",
                    "x-faker": "lorem.word"
                }
            },
            "required": [
                "value"
            ],
            "example": {
                "value": "1-047-149-5200 x41058",
                "type": "ut"
            }
        }
    }
}